analysis/requirements_analysis.txt
# analysis/requirements_analysis.txt

# Core data
pandas>=2.0
numpy>=1.26

# NLP / embeddings
sentence-transformers>=2.7
keybert>=0.8
langdetect>=1.0.0

# LLM provider SDKs
openai>=1.30
anthropic>=0.25
litellm>=1.40

# Approximate nearest neighbors
faiss-cpu>=1.8

# Graph
networkx>=3.3
python-louvain>=0.16
umap-learn>=0.5
pyvis>=0.3

# Visualization
plotly>=5.22
wordcloud>=1.9
matplotlib>=3.8

# Dashboard
streamlit>=1.35

# Utilities
pyyaml>=6.0
tqdm>=4.66
python-dotenv>=1.0
Levenshtein>=0.25

# Testing
pytest>=8.0
pytest-cov>=5.0
